+++ browser-extensions/chrome/src/ai-spacing/content-script.ts
